+++ test/sort/linear_complexity/radix_sort.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include "utils/tracer.hpp"
+#include "tracer_impl.hpp"
 #include "sort/linear_complexity/radix_sort.hpp"
 
 namespace sort
@@ -16,22 +16,24 @@ using RadixSortTestParams = std::tuple<array_t, digits_max_t, expected_array_t>;
 
 class RadixSortFixture : public testing::TestWithParam<RadixSortTestParams> {};
 
-
 namespace printer
 {
-    struct Tracer
+    struct RadixTracer : public Tracer
     {
+        using Tracer::operator ();
+        void operator() (size_t digit_num)
+        {
+            std::cout << "\nStart countingsort by: " << digit_num << " digit" << std::endl;
+        }
     };
 }
-
-
 TEST_P(RadixSortFixture, test)
 {
     array_t in_array = std::get<0>(GetParam());
     digits_max_t max_digits = std::get<1>(GetParam());
     expected_array_t expected = std::get<2>(GetParam());
 
-    printer::Tracer tracer;
+    printer::RadixTracer tracer;
     radix_sort(in_array, max_digits, tracer);
     ASSERT_EQ(in_array, expected) << "Unexpected sorted array";
 }
